+++ libs/utils/src/lib.rs
@@ -54,6 +54,8 @@ pub mod measured_stream;
 pub mod serde_percent;
 pub mod serde_regex;
 
+pub mod pageserver_feedback;
+
 pub mod tracing_span_assert;
 
 /// use with fail::cfg("$name", "return(2000)")
